wip: Fix test failure on windows
