+++ pkg/proc/proc_test.go
@@ -8,7 +8,6 @@ import (
 	"go/ast"
 	"go/constant"
 	"go/token"
-	"io"
 	"io/ioutil"
 	"math/rand"
 	"net"
@@ -21,7 +20,6 @@ import (
 	"sort"
 	"strconv"
 	"strings"
-	"syscall"
 	"testing"
 	"text/tabwriter"
 	"time"
@@ -45,82 +43,6 @@ var testBackend, buildMode string
 func init() {
 	runtime.GOMAXPROCS(4)
 	os.Setenv("GOMAXPROCS", "4")
-}
-
-func TestRedirect(t *testing.T) {
-	fixture := protest.BuildFixture("out_redirect", 0)
-	var (
-		p                *proc.Target
-		tracedir         string
-		err              error
-		redirect         proc.Redirect = proc.NewEmptyRedirectByPath()
-		errChan                        = make(chan error, 2)
-		canceFunc        func()
-		needCheck        = false
-		stdoutExpectFile = "out_redirect-stdout.txt"
-		stderrExpectFile = "out_redirect-stderr.txt"
-	)
-	switch testBackend {
-	case "native":
-		if runtime.GOOS == "linux" {
-			redirect, canceFunc, err = testGenRedireByPath(t, fixture, stdoutExpectFile, stderrExpectFile, errChan)
-			if err != nil {
-				break
-			}
-			needCheck = true
-		} else if runtime.GOOS == "windows" {
-			// TODO toad
-		}
-
-		p, err = native.Launch([]string{fixture.Path}, ".", 0, []string{}, "", redirect)
-	case "lldb":
-		if runtime.GOOS == "darwin" {
-			redirect, canceFunc, err = testGenRedireByPath(t, fixture, stdoutExpectFile, stderrExpectFile, errChan)
-			if err != nil {
-				break
-			}
-			needCheck = true
-		}
-		p, err = gdbserial.LLDBLaunch([]string{fixture.Path}, ".", 0, []string{}, "", redirect)
-	case "rr":
-		protest.MustHaveRecordingAllowed(t)
-		t.Log("recording")
-		if runtime.GOOS != "windows" {
-			redirect, canceFunc, err = testGenRedireByPath(t, fixture, stdoutExpectFile, stderrExpectFile, errChan)
-			if err != nil {
-				break
-			}
-			needCheck = true
-		}
-		p, tracedir, err = gdbserial.RecordAndReplay([]string{fixture.Path}, ".", true, []string{}, redirect)
-		t.Logf("replaying %q", tracedir)
-	default:
-		t.Fatal("unknown backend")
-	}
-	if err != nil {
-		// clear reader goroutine
-		if canceFunc != nil {
-			canceFunc()
-			_, _ = <-errChan, <-errChan
-		}
-		t.Fatal("Launch():", err)
-	}
-	_ = p.Continue()
-
-	if needCheck {
-		err1 := <-errChan
-		err2 := <-errChan
-		if err1 != nil {
-			t.Fatal("checkOut():", err1)
-		}
-		if err2 != nil {
-			t.Fatal("checkOut():", err2)
-		}
-	}
-
-	defer func() {
-		p.Detach(true)
-	}()
 }
 
 func TestMain(m *testing.M) {
@@ -6025,63 +5947,82 @@ func TestStacktraceExtlinkMac(t *testing.T) {
 	})
 }
 
-func testGenRedireByPath(t *testing.T, fixture protest.Fixture, stdoutExpectFile string, stderrExpectFile string, errChan chan error) (redirect proc.Redirect, canceFunc func(), err error) {
+func TestRedirect(t *testing.T) {
+	fixture := protest.BuildFixture("out_redirect", 0)
 	var (
-		stdoutPath = "./stdout"
-		stderrPath = "./stderr"
+		p                *proc.Target
+		tracedir         string
+		err              error
+		redirect         proc.Redirect = proc.NewEmptyRedirectByPath()
+		errChan                        = make(chan error, 2)
+		canceFunc        func()
+		needCheck        = false
+		stdoutExpectFile = "out_redirect-stdout.txt"
+		stderrExpectFile = "out_redirect-stderr.txt"
 	)
+	switch testBackend {
+	case "native":
+		if runtime.GOOS == "linux" {
+			redirect, canceFunc, err = testGenRedireByPath(t, fixture, stdoutExpectFile, stderrExpectFile, errChan)
+			if err != nil {
+				break
+			}
+			needCheck = true
+		} else if runtime.GOOS == "windows" {
+			redirect, canceFunc, err = testGenRedireByFile(t, fixture, stdoutExpectFile, stderrExpectFile, errChan)
+			if err != nil {
+				break
+			}
+			needCheck = true
+		}
 
-	if err = syscall.Mkfifo(stdoutPath, 0o600); err != nil {
-		return redirect, nil, err
+		p, err = native.Launch([]string{fixture.Path}, ".", 0, []string{}, "", redirect)
+	case "lldb":
+		if runtime.GOOS == "darwin" {
+			redirect, canceFunc, err = testGenRedireByPath(t, fixture, stdoutExpectFile, stderrExpectFile, errChan)
+			if err != nil {
+				break
+			}
+			needCheck = true
+		}
+		p, err = gdbserial.LLDBLaunch([]string{fixture.Path}, ".", 0, []string{}, "", redirect)
+	case "rr":
+		protest.MustHaveRecordingAllowed(t)
+		t.Log("recording")
+		if runtime.GOOS != "windows" {
+			redirect, canceFunc, err = testGenRedireByPath(t, fixture, stdoutExpectFile, stderrExpectFile, errChan)
+			if err != nil {
+				break
+			}
+			needCheck = true
+		}
+		p, tracedir, err = gdbserial.RecordAndReplay([]string{fixture.Path}, ".", true, []string{}, redirect)
+		t.Logf("replaying %q", tracedir)
+	default:
+		t.Fatal("unknown backend")
+	}
+	if err != nil {
+		// clear reader goroutine
+		if canceFunc != nil {
+			canceFunc()
+			_, _ = <-errChan, <-errChan
+		}
+		t.Fatal("Launch():", err)
+	}
+	_ = p.Continue()
+
+	if needCheck {
+		err1 := <-errChan
+		err2 := <-errChan
+		if err1 != nil {
+			t.Fatal("checkOut():", err1)
+		}
+		if err2 != nil {
+			t.Fatal("checkOut():", err2)
+		}
 	}
 
-	if err = syscall.Mkfifo(stderrPath, 0o600); err != nil {
-		os.Remove(stdoutPath)
-		return redirect, nil, err
-	}
-
-	redirect = proc.NewRedirectByPath([3]string{"", stdoutPath, stderrPath})
-	reader := func(mode string, path string, expectFile string) {
-		defer os.Remove(path)
-		fmt.Println(path)
-		outFile, err := os.OpenFile(path, os.O_RDONLY, os.ModeNamedPipe)
-		if err != nil {
-			errChan <- err
-			return
-		}
-
-		expect, err := os.ReadFile(filepath.Join(fixture.BuildDir, expectFile))
-		if err != nil {
-			errChan <- err
-			return
-		}
-
-		out, err := io.ReadAll(outFile)
-		if err != nil {
-			errChan <- err
-			return
-		}
-
-		if string(expect) == string(out) {
-			errChan <- fmt.Errorf("%s,Not as expected!\nexpect:%s\nout:%s", mode, expect, out)
-			return
-		}
-		errChan <- nil
-	}
-
-	canceFunc = func() {
-		stdout, err := os.OpenFile(stdoutPath, os.O_WRONLY, os.ModeNamedPipe)
-		if err == nil {
-			stdout.Close()
-		}
-		stderr, err := os.OpenFile(stderrPath, os.O_WRONLY, os.ModeNamedPipe)
-		if err == nil {
-			stderr.Close()
-		}
-	}
-
-	go reader("stdout", stdoutPath, stdoutExpectFile)
-	go reader("stderr", stderrPath, stderrExpectFile)
-
-	return redirect, canceFunc, nil
+	defer func() {
+		p.Detach(true)
+	}()
 }
